revert parser changes
